refactor: moved observers to extensions folder
