--- src/domain/observers/GithubAction.ts
@@ -1,59 +1,0 @@
-import { Octokit } from '@octokit/rest';
-import { Observer } from '@/src/types/Observer';
-import { State } from '@/src/types/State';
-import { GithubActionConfiguration } from '@/src/types/GithubActionConfiguration';
-import { Status } from '@/src/types/Status';
-
-export class GithubAction implements Observer {
-  private octokit: Octokit;
-  private owner: string;
-  private repo: string;
-  private workflowId: string;
-
-  private alias: string;
-  private baseLink: string;
-
-  constructor({ authToken, owner, repo, workflowId, alias }: GithubActionConfiguration) {
-    this.octokit = new Octokit({
-      auth: authToken,
-    });
-    this.owner = owner;
-    this.repo = repo;
-    this.workflowId = workflowId;
-    this.alias = alias || `Github: ${this.owner}/${this.repo}/${this.workflowId}`;
-    this.baseLink = `https://github.com/${owner}/${repo}/actions/`;
-  }
-
-  public async getState(): Promise<State> {
-    try {
-      const response = await this.octokit.request('GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs', {
-        owner: this.owner,
-        repo: this.repo,
-        workflow_id: this.workflowId,
-      });
-      if (response.status != 200 || response.data.total_count == 0) throw new Error('response is invalid');
-      const { conclusion, html_url } = response.data.workflow_runs[0];
-      return {
-        name: this.alias,
-        status: this.getStatus(conclusion),
-        link: html_url,
-      };
-      return {
-        name: this.alias,
-        status: Status.NA,
-        link: this.baseLink,
-      };
-    } catch (error) {
-      console.error(error);
-      return {
-        name: this.alias,
-        status: Status.NA,
-        link: this.baseLink,
-      };
-    }
-  }
-
-  private getStatus(conclusion: string | null): Status {
-    return conclusion == null ? Status.CHECKING : conclusion === 'success' ? Status.SUCCESS : Status.FAILURE;
-  }
-}
