+++ src/extensions/github/observer/index.ts
@@ -1,8 +1,8 @@
-import { Octokit } from 'octokit';
-import { Observer } from '../../../types/Observer';
-import { State } from '../../../types/State';
-import { GithubActionConfiguration } from '../../../types/GithubActionConfiguration';
-import { Status } from '../../../types/Status';
+import { Octokit } from '@octokit/rest';
+import { Observer } from '@/src/types/Observer';
+import { State } from '@/src/types/State';
+import { GithubActionConfiguration } from '@/src/types/GithubActionConfiguration';
+import { Status } from '@/src/types/Status';
 
 export class GithubAction implements Observer {
   private octokit: Octokit;
@@ -48,7 +48,7 @@ export class GithubAction implements Observer {
     }
   }
 
-  private getStatus(conclusion: string): Status {
+  private getStatus(conclusion: string | null): Status {
     return conclusion == null ? Status.CHECKING : conclusion === 'success' ? Status.SUCCESS : Status.FAILURE;
   }
 }
