+++ src/types/CCTrayConfiguration.ts
@@ -2,5 +2,5 @@ import { ObserverConfiguration } from '@/src/types/ObserverConfiguration';
 
 export type CCTrayConfiguration = ObserverConfiguration & {
   url: string;
-  name?: string;
+  name: string;
 };
